создал модуль thumbnails.js отрисовал миниатюры
