+++ js/data.js
@@ -1,58 +1,8 @@
 // Модуль с данными для фотографий
-
 import { getRandomArrayElement, getRandomInteger } from './util';
+import {DESCRIPTIONS, MESSAGES, NAMES, } from './constants.js';
 
-const MAX_PHOTO_POSTS = 25;
-
-const DESCRIPTIONS = [
-  'Будь таким человеком, с которым мечтаешь встретиться',
-  'Будьте героями своих собственных историй',
-  'В жизни я лучше, чем в инстаграмме',
-  'В любой ситуации всегда улыбайтесь',
-  'Думать следует до и после съемки, ни никотгда во время нее',
-  'В своей жизни ты должен играть только главную роль',
-  'Ваша скорость не имеет значения, пока вы продолжаете двигаться вперед',
-  'На всяукий случай, а то вдруг вы забыли как я выгляжу',
-  'Время драгоценно, поэтому всегда тратьте его мудро'
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Jacob',
-  'Oliver',
-  'Riley',
-  'Jack',
-  'Alfie',
-  'Harry',
-  'Charlie',
-  'Dylan',
-  'William',
-  'Mason',
-  'Amelia',
-  'Ava',
-  'Mia',
-  'Lily',
-  'Olivia',
-  'Ruby',
-  'Seren',
-  'Evie',
-  'Ella',
-  'Emily',
-  'Igor',
-  'Andrew',
-  'Pavel',
-  'Nataly',
-  'Alexandra'
-];
-
+let photoCount;
 //Функция для создания обьекта с описанием фотографии
 const createPhotoPost = () => {
   const commentsCount = getRandomInteger(0, 30);
@@ -80,7 +30,7 @@ const createComment = () => {
 const generateId = () => {
   let id = 0;
   return function () {
-    while (id < MAX_PHOTO_POSTS) {
+    while (id < photoCount) {
       id += 1;
       return id;
     }
@@ -99,6 +49,11 @@ const createCommentId = generateCommentId();
 
 const createId = generateId();
 
-const createPhotoPostArray = () => Array.from({ length: MAX_PHOTO_POSTS }, createPhotoPost);
+const createPhotoPostArray = (n) => {
+  photoCount = n;
+  return Array.from({ length: photoCount }, createPhotoPost);
+
+}
+
 
 export { createPhotoPostArray };
